Start making filters sidebar sticky and move prepareDatasets logic
