+++ app/scripts/components/data-catalog/filters-control.tsx
@@ -7,7 +7,10 @@ import { Actions, useBrowserControls } from '$components/common/browse-controls/
 
 const ControlsWrapper = styled.div<{ width?: string; }>`
   min-width: 20rem;
-  width: ${props => props.width ?? '3rem'};
+  width: ${props => props.width ?? '20rem'};
+  position: sticky;
+  top: 0;
+  height: 100vh;
 `;
 
 interface FiltersMenuProps extends ReturnType<typeof useBrowserControls> {
@@ -42,6 +45,10 @@ export default function FiltersControl(props: FiltersMenuProps) {
     }
   }, [allSelected, setClearedTagItem, onChangeToFilters]);
 
+
+  const displayFiltersOpened = React.useMemo(() => taxonomiesOptions.length > 4 ? false : true, [taxonomiesOptions]);
+
+
   return (
     <ControlsWrapper width={width}>
       <SearchField
@@ -61,6 +68,7 @@ export default function FiltersControl(props: FiltersMenuProps) {
               onChanges={handleChanges}
               globallySelected={allSelected}
               tagItemCleared={{item: clearedTagItem, callback: setClearedTagItem}}
+              showFiltersOpened={displayFiltersOpened}
             />
           );
         })
