+++ app/scripts/components/data-catalog/index.tsx
@@ -1,6 +1,6 @@
 import React, { useRef } from 'react';
 import styled from 'styled-components';
-import { DatasetData, getString } from 'veda';
+import { DatasetData } from 'veda';
 import { Link, useNavigate } from 'react-router-dom';
 import { themeVal } from '@devseed-ui/theme-provider';
 import { VerticalDivider } from '@devseed-ui/toolbar';
@@ -8,16 +8,15 @@ import { VerticalDivider } from '@devseed-ui/toolbar';
 import DatasetMenu from './dataset-menu';
 import FiltersControl from './filters-control';
 import FilterTag from './filter-tag';
+import prepareDatasets from './prepare-datasets';
 import {
   Actions,
-  optionAll,
   useBrowserControls
 } from '$components/common/browse-controls/use-browse-controls';
 import {
-  LayoutProps,
   useSlidingStickyHeaderProps
 } from '$components/common/layout-root';
-import PageHero from '$components/common/page-hero';
+
 import {
   FoldHeader,
   FoldHeadline,
@@ -26,11 +25,9 @@ import {
 import { Card } from '$components/common/card';
 import { CardList, CardMeta, CardTopicsList } from '$components/common/card/styles';
 import EmptyHub from '$components/common/empty-hub';
-import { PageMainContent } from '$styles/page';
 import { DATASETS_PATH, getDatasetPath } from '$utils/routes';
 import TextHighlight from '$components/common/text-highlight';
 import { Pill } from '$styles/pill';
-import { FeaturedDatasets } from '$components/common/featured-slider-section';
 import { CardSourcesList } from '$components/common/card-sources';
 import {
   getAllTaxonomyValues,
@@ -46,6 +43,11 @@ import { OptionItem } from '$components/common/form/checkable-filter';
 import { usePreviousValue } from '$utils/use-effect-previous';
 import { getAllDatasetsWithEnhancedLayers } from '$components/exploration/data-utils';
 
+/**
+ * DATA CATALOG Feature component 
+ * Allows you to browse through datasets using the filters sidebar control
+ */
+
 const BrowseFoldHeader = styled(FoldHeader)`
   margin-bottom: 4rem;
 `;
@@ -53,6 +55,7 @@ const BrowseFoldHeader = styled(FoldHeader)`
 const Content = styled.div`
   display: flex;
   margin-bottom: 8rem;
+  position: sticky;
 `;
 
 const CatalogWrapper = styled.div`
@@ -68,6 +71,10 @@ const BrowseSection = styled.div`
   padding-right: ${variableGlsp()};
   gap: ${variableGlsp()};
 `;
+
+// const FilterWrapper = styled.div`
+//   display: block;
+// `;
 
 const Cards = styled(CardList)`
   padding: 0 0 0 2rem;
@@ -97,83 +104,6 @@ const EmptyState = styled(EmptyHub)`
 `;
 
 export const sortOptions = [{ id: 'name', name: 'Name' }];
-
-export const prepareDatasets = (
-  data: DatasetData[],
-  options: {
-    search: string;
-    taxonomies: Record<string, string | string[]> | null;
-    sortField: string | null;
-    sortDir: string | null;
-    filterLayers: boolean | null;
-  }
-) => {
-  const { sortField, sortDir, search, taxonomies, filterLayers } = options;
-  let filtered = [...data];
-
-  // Does the free text search appear in specific fields?
-  if (search.length >= 3) {
-    const searchLower = search.toLowerCase();
-    // Function to check if searchLower is included in any of the string fields
-    const includesSearchLower = (str) => str.toLowerCase().includes(searchLower);
-    // Function to determine if a layer matches the search criteria
-    const layerMatchesSearch = (layer) => 
-      includesSearchLower(layer.stacCol) ||
-      includesSearchLower(layer.name) ||
-      includesSearchLower(layer.parentDataset.name) ||
-      includesSearchLower(layer.parentDataset.id) ||
-      includesSearchLower(layer.description);
-
-    filtered = filtered
-      .filter((d) => {
-        // Pre-calculate lowercased versions to use in comparisons
-        const idLower = d.id.toLowerCase();
-        const nameLower = d.name.toLowerCase();
-        const descriptionLower = d.description.toLowerCase();
-        const topicsTaxonomy = d.taxonomy.find((t) => t.name === TAXONOMY_TOPICS);
-        // Check if any of the conditions for including the item are met
-        return (
-          idLower.includes(searchLower) ||
-          nameLower.includes(searchLower) ||
-          descriptionLower.includes(searchLower) ||
-          d.layers.some(layerMatchesSearch) ||
-          topicsTaxonomy?.values.some((t) => includesSearchLower(t.name))
-        );
-      });
-
-      if (filterLayers)
-        filtered = filtered.map((d) => ({
-          ...d,
-          layers: d.layers.filter(layerMatchesSearch),
-        }));
-  }
-
-  taxonomies &&
-    Object.entries(taxonomies).forEach(([name, value]) => {
-      if (!value.includes(optionAll.id)) {
-        filtered = filtered.filter((d) =>
-          d.taxonomy.some(
-            (t) => t.name === name && t.values.some((v) => value.includes(v.id))
-          )
-        );
-      }
-    });
-
-  sortField &&
-    /* eslint-disable-next-line fp/no-mutating-methods */
-    filtered.sort((a, b) => {
-      if (!a[sortField]) return Infinity;
-
-      return a[sortField]?.localeCompare(b[sortField]);
-    });
-
-  if (sortDir === 'desc') {
-    /* eslint-disable-next-line fp/no-mutating-methods */
-    filtered.reverse();
-  }
-
-  return filtered;
-};
 
 export interface DataCatalogProps {
   datasets: DatasetData[];
@@ -212,6 +142,10 @@ function DataCatalog({ datasets }: DataCatalogProps) {
 
   const prevSelectedFilters = usePreviousValue(allSelectedFilters) || [];
 
+  // const targetRef = React.useRef<HTMLDivElement>(null);
+  // const [targetHeight, setTargetHeight] = React.useState<number>(0);
+
+  // Handlers
   const handleChangeAllSelectedFilters = React.useCallback((item: OptionItem, action: 'add' | 'remove') => {
     if(action == 'add') {
       setAllSelectedFilters([...allSelectedFilters, item]);
@@ -232,6 +166,13 @@ function DataCatalog({ datasets }: DataCatalogProps) {
   const handleClearTags = React.useCallback(() => {
     setAllSelectedFilters([]);
   }, [setAllSelectedFilters]);
+
+  // React.useEffect(() => {
+  //   if(targetRef.current) {
+  //     const height = targetRef.current.offsetHeight;
+  //     console.log(`height: `, height);
+  //   }
+  // }, [targetRef]);
 
   React.useEffect(() => {
     if (clearedTagItem && (allSelectedFilters.length == prevSelectedFilters.length-1)) {
@@ -280,29 +221,19 @@ function DataCatalog({ datasets }: DataCatalogProps) {
   }, [allSelectedFilters, handleClearTag, handleClearTags, urlTaxonomyItems]);
 
   return (
-    <PageMainContent>
-      <LayoutProps
-        title='Data Catalog'
-        description={getString('dataCatalogBanner').other}
-      />
-      <PageHero
-        title='Data Catalog'
-        description={getString('dataCatalogBanner').other}
-      />
-
-      <FeaturedDatasets />
-      <BrowseSection>
-        <BrowseFoldHeader
-          ref={browseControlsHeaderRef}
-          style={{
-            scrollMarginTop: `${headerHeight + 16}px`
-          }}
-        >
-          <FoldHeadline>
-            <FoldTitle>Search datasets</FoldTitle>
-          </FoldHeadline>
-        </BrowseFoldHeader>
-        <Content>
+    <BrowseSection>
+      <BrowseFoldHeader
+        ref={browseControlsHeaderRef}
+        style={{
+          scrollMarginTop: `${headerHeight + 16}px`
+        }}
+      >
+        <FoldHeadline>
+          <FoldTitle>Search datasets</FoldTitle>
+        </FoldHeadline>
+      </BrowseFoldHeader>
+      <Content>
+        {/* <FilterWrapper ref={targetRef}> */}
           <FiltersControl
             {...controlVars}
             taxonomiesOptions={datasetTaxonomies}
@@ -311,121 +242,121 @@ function DataCatalog({ datasets }: DataCatalogProps) {
             setClearedTagItem={setClearedTagItem}
             allSelected={allSelectedFilters}
           />
-          <CatalogWrapper>
-            {renderTags}
-            {datasetsToDisplay.length ? (
-              <Cards>
-                {datasetsToDisplay.map((d) => {
-                  const topics = getTaxonomy(d, TAXONOMY_TOPICS)?.values;
-                  const allTaxonomyValues = getAllTaxonomyValues(d).map((v) => v.name);
-                  return (
-                    <li key={d.id}>
-                      <Card
-                        cardType='horizontal-info'
-                        tagLabels={allTaxonomyValues}
-                        overline={
-                          <CardMeta>
-                            <DatasetClassification dataset={d} />
-                            <CardSourcesList
-                              sources={getTaxonomy(d, TAXONOMY_SOURCE)?.values}
-                              rootPath={DATASETS_PATH}
-                              onSourceClick={(id) => {
-                                onAction(Actions.TAXONOMY_MULTISELECT, {
-                                  key: TAXONOMY_SOURCE,
-                                  value: id
-                                });
-                                browseControlsHeaderRef.current?.scrollIntoView();
-                              }}
-                            />
-                            <VerticalDivider variation='light' />
-                            {/* TODO: Implement modified date: https://github.com/NASA-IMPACT/veda-ui/issues/514 */}
-                            {/* 
-                          <Link
-                            to={`${DATASETS_PATH}?${Actions.SORT_FIELD}=date`}
-                            onClick={(e) => {
-                              e.preventDefault();
-                              onAction(Actions.SORT_FIELD, 'date');
+        {/* </FilterWrapper> */}
+        <CatalogWrapper>
+          {renderTags}
+          {datasetsToDisplay.length ? (
+            <Cards>
+              {datasetsToDisplay.map((d) => {
+                const topics = getTaxonomy(d, TAXONOMY_TOPICS)?.values;
+                const allTaxonomyValues = getAllTaxonomyValues(d).map((v) => v.name);
+                return (
+                  <li key={d.id}>
+                    <Card
+                      cardType='horizontal-info'
+                      tagLabels={allTaxonomyValues}
+                      overline={
+                        <CardMeta>
+                          <DatasetClassification dataset={d} />
+                          <CardSourcesList
+                            sources={getTaxonomy(d, TAXONOMY_SOURCE)?.values}
+                            rootPath={DATASETS_PATH}
+                            onSourceClick={(id) => {
+                              onAction(Actions.TAXONOMY_MULTISELECT, {
+                                key: TAXONOMY_SOURCE,
+                                value: id
+                              });
+                              browseControlsHeaderRef.current?.scrollIntoView();
                             }}
-                          >
-                            Updated <time dateTime='2023-01-01'>X time ago</time>
-                          </Link> */}
-                          </CardMeta>
-                        }
-                        linkLabel='View more'
-                        linkTo={getDatasetPath(d)}
-                        title={
-                          <TextHighlight
-                            value={search}
-                            disabled={search.length < 3}
-                          >
-                            {d.name}
-                          </TextHighlight>
-                        }
-                        description={
-                          <TextHighlight
-                            value={search}
-                            disabled={search.length < 3}
-                          >
-                            {d.description}
-                          </TextHighlight>
-                        }
-                        imgSrc={d.media?.src}
-                        imgAlt={d.media?.alt}
-                        footerContent={
-                          <>
-                            {topics?.length ? (
-                              <CardTopicsList>
-                                <dt>Topics</dt>
-                                {topics.map((t) => {
-                                  const path = `${DATASETS_PATH}?${
-                                    Actions.TAXONOMY
-                                  }=${encodeURIComponent(
-                                    JSON.stringify({ Topics: [t.id] })
-                                  )}`;
-                                  return (
-                                    <dd key={t.id}>
-                                      <Pill
-                                        variation='achromic'
-                                        as={Link}
-                                        to={path}
-                                        onClick={(e) => {
-                                          e.preventDefault();
-                                          onAction(Actions.TAXONOMY_MULTISELECT, {
-                                            key: TAXONOMY_TOPICS,
-                                            value: t.id
-                                          });
-                                          browseControlsHeaderRef.current?.scrollIntoView();
-                                        }}
+                          />
+                          <VerticalDivider variation='light' />
+                          {/* TODO: Implement modified date: https://github.com/NASA-IMPACT/veda-ui/issues/514 */}
+                          {/* 
+                        <Link
+                          to={`${DATASETS_PATH}?${Actions.SORT_FIELD}=date`}
+                          onClick={(e) => {
+                            e.preventDefault();
+                            onAction(Actions.SORT_FIELD, 'date');
+                          }}
+                        >
+                          Updated <time dateTime='2023-01-01'>X time ago</time>
+                        </Link> */}
+                        </CardMeta>
+                      }
+                      linkLabel='View more'
+                      linkTo={getDatasetPath(d)}
+                      title={
+                        <TextHighlight
+                          value={search}
+                          disabled={search.length < 3}
+                        >
+                          {d.name}
+                        </TextHighlight>
+                      }
+                      description={
+                        <TextHighlight
+                          value={search}
+                          disabled={search.length < 3}
+                        >
+                          {d.description}
+                        </TextHighlight>
+                      }
+                      imgSrc={d.media?.src}
+                      imgAlt={d.media?.alt}
+                      footerContent={
+                        <>
+                          {topics?.length ? (
+                            <CardTopicsList>
+                              <dt>Topics</dt>
+                              {topics.map((t) => {
+                                const path = `${DATASETS_PATH}?${
+                                  Actions.TAXONOMY
+                                }=${encodeURIComponent(
+                                  JSON.stringify({ Topics: [t.id] })
+                                )}`;
+                                return (
+                                  <dd key={t.id}>
+                                    <Pill
+                                      variation='achromic'
+                                      as={Link}
+                                      to={path}
+                                      onClick={(e) => {
+                                        e.preventDefault();
+                                        onAction(Actions.TAXONOMY_MULTISELECT, {
+                                          key: TAXONOMY_TOPICS,
+                                          value: t.id
+                                        });
+                                        browseControlsHeaderRef.current?.scrollIntoView();
+                                      }}
+                                    >
+                                      <TextHighlight
+                                        value={search}
+                                        disabled={search.length < 3}
                                       >
-                                        <TextHighlight
-                                          value={search}
-                                          disabled={search.length < 3}
-                                        >
-                                          {t.name}
-                                        </TextHighlight>
-                                      </Pill>
-                                    </dd>
-                                  );
-                                })}
-                              </CardTopicsList>
-                            ) : null}
-                            <DatasetMenu dataset={d} />
-                          </>
-                        }
-                      />
-                    </li>
-                  );
-                })}
-              </Cards>
-            ) : (
-              <EmptyState>
-                There are no datasets to show with the selected filters.
-              </EmptyState>
-            )}
-          </CatalogWrapper>
-        </Content>
-      </BrowseSection>
-    </PageMainContent>
+                                        {t.name}
+                                      </TextHighlight>
+                                    </Pill>
+                                  </dd>
+                                );
+                              })}
+                            </CardTopicsList>
+                          ) : null}
+                          <DatasetMenu dataset={d} />
+                        </>
+                      }
+                    />
+                  </li>
+                );
+              })}
+            </Cards>
+          ) : (
+            <EmptyState>
+              There are no datasets to show with the selected filters.
+            </EmptyState>
+          )}
+        </CatalogWrapper>
+      </Content>
+    </BrowseSection>
   );
 }
