+++ app/scripts/components/exploration/components/dataset-selector-modal/index.tsx
@@ -28,7 +28,8 @@ import {
   TaxonomyFilterOption,
   useBrowserControls
 } from '$components/common/browse-controls/use-browse-controls';
-import { prepareDatasets, sortOptions } from '$components/data-catalog';
+import { sortOptions } from '$components/data-catalog';
+import prepareDatasets from '$components/data-catalog/prepare-datasets';
 import { TAXONOMY_SOURCE, getTaxonomy } from '$utils/veda-data';
 import { usePreviousValue } from '$utils/use-effect-previous';
